Clean up a ton of Caliper warnings. ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=29725661

+++ examples/src/main/java/examples/CopyArrayBenchmark.java
@@ -271,7 +271,7 @@ public class CopyArrayBenchmark extends SimpleBenchmark {
       byteArray[i] = (byte) num;
       charArray[i] = (char) num;
       doubleArray[i] = num;
-      floatArray[i] = (float) num;
+      floatArray[i] = num;
       intArray[i] = num;
       longArray[i] = num;
       shortArray[i] = (short) num;
